♻️ Refactor: 타입 분리
refactor: 타입 분리
