+++ src/features/d-day/model/state.ts
@@ -1,0 +1,8 @@
+import { Emoji } from "@/entities/d-day";
+
+export interface InitialDday {
+  id: number | null;
+  title: string;
+  date: string;
+  emoji: Emoji;
+}
